Move calendar to activity
