+++ app/src/main/res/layout/fragment_five_things.xml
@@ -74,68 +74,7 @@
                     app:layout_constraintTop_toTopOf="parent"
                     tools:text="Wednesday September 24th, 2018" />
 
-                <android.support.constraint.ConstraintLayout
-                    android:id="@+id/calendar"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginEnd="20dp"
-                    android:layout_marginStart="20dp"
-                    android:layout_marginTop="8dp"
-                    android:background="@drawable/rounded_button"
-                    android:visibility="@{calendarVisible ?  View.VISIBLE : View.GONE}"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@+id/current_date"
-                    tools:visibility="gone">
 
-                    <TextView
-                        android:id="@+id/todayButton"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginEnd="20dp"
-                        android:layout_marginTop="10dp"
-                        android:text="Today"
-                        android:textColor="@color/white"
-                        app:layout_constraintBottom_toBottomOf="@+id/month_year"
-                        app:layout_constraintEnd_toEndOf="parent"
-                        app:layout_constraintTop_toTopOf="parent" />
-
-                    <TextView
-                        android:id="@+id/month_year"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_marginStart="20dp"
-                        android:layout_marginTop="8dp"
-                        android:gravity="center_horizontal"
-                        android:text="@{month}"
-                        android:textColor="@color/white"
-                        android:textSize="20sp"
-                        app:layout_constraintStart_toStartOf="parent"
-                        app:layout_constraintTop_toTopOf="parent"
-                        tools:text="Month YEAR" />
-
-                    <com.github.sundeepk.compactcalendarview.CompactCalendarView
-                        android:id="@+id/compactcalendar_view"
-                        android:layout_width="match_parent"
-                        android:layout_height="250dp"
-                        android:layout_marginBottom="8dp"
-                        android:layout_marginLeft="8dp"
-                        android:layout_marginRight="8dp"
-                        android:paddingLeft="10dp"
-                        android:paddingRight="10dp"
-                        app:compactCalendarBackgroundColor="@color/bluegreen"
-                        app:compactCalendarCurrentDayBackgroundColor="@color/another_grey"
-                        app:compactCalendarCurrentSelectedDayBackgroundColor="@color/grey"
-                        app:compactCalendarMultiEventIndicatorColor="@color/white"
-                        app:compactCalendarTargetHeight="250dp"
-                        app:compactCalendarTextColor="@color/white"
-                        app:compactCalendarTextSize="12sp"
-                        app:layout_constraintBottom_toBottomOf="parent"
-                        app:layout_constraintEnd_toEndOf="parent"
-                        app:layout_constraintHorizontal_bias="1.0"
-                        app:layout_constraintStart_toStartOf="parent"
-                        app:layout_constraintTop_toBottomOf="@+id/month_year" />
-                </android.support.constraint.ConstraintLayout>
 
                 <TextView
                     android:id="@+id/one_label"
